webapi/webapi-webrtc-w3c-tests/webrtc/RTCPeerConnection_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>WebRTC Test: RTCPeerConnection basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcpeerconnection-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var pc;

setup(function () {
  var configuration = {"iceServers": [{"url": "stun:stun.example.org"}]};
  if (typeof RTCPeerConnection != "undefined") {
    pc = new RTCPeerConnection(configuration);
  } else {
    pc = new webkitRTCPeerConnection(configuration);
  }
});

[
  //attributes of the RTCPeerConnection interface
  ["object", "localDescription", "readonly"],
  ["object", "remoteDescription", "readonly"],
  ["string", "signalingState", "readonly"],
  ["string", "iceGatheringState", "readonly"],
  ["string", "iceConnectionState", "readonly"],

  //events of the RTCPeerConnection interface
  ["event", "onnegotiationneeded", ""],
  ["event", "onicecandidate", ""],
  ["event", "onsignalingstatechange", ""],
  ["event", "onaddstream", ""],
  ["event", "onremovestream", ""],
  ["event", "oniceconnectionstatechange", ""],
  ["event", "ondatachannel", ""],

  //methods of the RTCPeerConnection interface
  ["function", "createOffer", ""],
  ["function", "createAnswer", ""],
  ["function", "setLocalDescription", ""],
  ["function", "setRemoteDescription", ""],
  ["function", "updateIce", ""],
  ["function", "addIceCandidate", ""],
  ["function", "getConfiguration", ""],
  ["function", "getLocalStreams", ""],
  ["function", "getRemoteStreams", ""],
  ["function", "getStreamById", ""],
  ["function", "addStream", ""],
  ["function", "removeStream", ""],
  ["function", "close", ""],
  ["function", "createDataChannel", ""],
  ["function", "getStats", ""],

  //the Identity API extends the RTCPeerConnection interface
  ["rtcIdentityAssertion", "peerIdentity", "readonly"],
  ["event", "onidentityresult", ""],
  ["event", "onpeeridentity", ""],
  ["event", "onidpassertionerror", ""],
  ["event", "onidpvalidationerror", ""],
  ["function", "setIdentityProvider", ""],
  ["function", "getIdentityAssertion", ""]
].forEach(function(attr) {
  var type = attr[0];
  var name = attr[1];
  var read = attr[2];

  test(function () {
    assert_true(name in pc, name + " attribute in RTCPeerConnection");
    switch(type) {
    case "string":
    case "object":
      if(read == "readonly") {
        assert_readonly(pc, name, "expect attribute " + name + " cannot be changed but");
      }
    case "function":
      assert_equals(typeof pc[name], type, name + " attribute of type");      
      break;
    case "rtcIdentityAssertion":
      assert_equals(Object.prototype.toString.call(pc[name]), "[object RTCIdentityAssertion]", name + " attribute of type");
      assert_readonly(pc, name, "expect attribute " + name + " cannot be changed but");
      break;
    default:
      break;
    }
  }, "Check if " + read + " RTCPeerConnection." + name + " exists and type of " + type);
});

</script>
